Add ability to run the querier without rendering
I'm rendering with my own script so i dont need it to be on. But i do
want to be capturing new renders. This will convert my known images and
allow me to keep tracking new stuff so I'm building up a huge base. It
might take longer to get all my images rendered and for me to decide to
start the stream, but this is cool