random_tests.py: Add test to ensure db can unicode
Phase01 script was failing when prompts included non-ascii unicode
characters. I found the root cause was the CHARACTER SET configured on
the table

The answer was to DROP the database and recreate it specifying the
character set so all tables in it would default to that character set.
The command for that were
DROP DATABASE wcddb;
CREATE DATABASE wcddb CHARACTER SET = 'utf8mb4'

I found out this was the answer because running this command on my
existing table fixed the problem
ALTER TABLE prompt CONVERT to CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci