Add script for the 24/7 fallback stream
This script, when running 24/7 and pointing to a YouTube broadcast will
keep that broadcast alive forever (tested with 3 hours so far). It also
has an intentional interruption injected so it will always only be the
backup. Because it stops sending data every minute or so if there is a
more reliable stream running (the primary) it will take priority