Tweak fallback stream for to me up for longer
With the old parameters the fallback stream (running by itself) dropped
the connection after 6h50m. With these new parameters, it has been
running on its own for 15+ hours. Also it appears to still defer to the
primary stream when one is running (at least it does defer after a few
minutes)