Phase04: Commit improved stream settings
I tweaked some stream settings to improve the quality of the stream
I'm not sure which ones made the biggest difference. So I should
probably take this apart and try each of these for now. But for now we
have a setting that streams smooth, stable, and high enough quality