Phase 3: Check for 404 codes from failed images
I noticed that phase 3 was crashing around 22 times in the first 20
something hours of running. The error case was caused by images where
the final_url cant be downloaded. So I pulled the download code out into
a function, wrote a test for it, and, modified phase 3 to skip failures
and not get stuck looking on them

For crash bugs in phase 3 they tend to repeat themselves until the thing
falls out of the top n. I should probably figure out how to prevent this
kind of retry loop