Phase 04: Refactor to track filesystem not clocks
With tha addition of fuser as a depenency that can reliably asses which
file ffmpeg is reading from, we can track the files as they go out. This
eliminates the need to track clock times and "predict" what video is
being shared. Instead we can just queue up a video when ffmpeg finishes
one

The real reason this was necessary is because I observed that ffmpeg
would get in a weird state somewhere around 34 hours into running. I
assumed it was because we were overwriting a file while ffmpeg was
reading it because of clock drift. This new error case is now impossible