Phase3: Tweak video timing now that lengths can vary
Before I had the insight into ffmpeg that fuser provides, we needed to
rely on the assumption that all videos were 20s. Now we can make them
dynamic. For now I set it to show all render stages, final image for 2s
and final image with annotations for 5s

Adjustments can be made in the future