Phase3: Performance and reliability tweaks
If we render videos with "fast" we can increase throughput and decrease
load on CPU, it seems. This was necessary because i think phase 3 was
starting to fall behind phase4

I also made the video cleanup step a little less agressive, it will
allow a large bank of 50 "ready" videos to stick around. This should
hopefully reduce the number of crashes i have in phase04 resulting from
output files being not found at the time they are to be moved into the
video queue