phase03: Fix crash bug by ignoring empty 0% images
Renderstages where percentage=0, if queried are most often empty images.
This was causing phase03 to crash and for us to get repeated images on
the stream. To fix this is just set phase03 to ignore those empty images

I probably should add some try/catch the the rendering code so that any
bad images dont cause the stream to backup, but this is fine for now